Optimized TextMesh.Append: although just a little effect. Change underlying type of Vector to float.

+++ src/ImGui/OSImplentation/TypographyTextContext.cs
@@ -45,7 +45,7 @@ namespace ImGui.OSImplentation
             return glyph;
         }
 
-        public static double GetScale(string fontFamily, double fontSize)
+        public static float GetScale(string fontFamily, double fontSize)
         {
             Typeface typeFace = GetTypeFace(fontFamily);
             var scale = typeFace.CalculateToPixelScaleFromPointSize((float)fontSize);
